simplify test files
